pyproject.toml
[project]
name = "pokemon-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.4",
    "beautifulsoup4>=4.13.4",
    "fastapi[all]>=0.115.12",
    "google-generativeai>=0.8.5",
    "html2text>=2025.4.15",
    "mcp[cli]>=1.9.1",
    "nest-asyncio>=1.6.0",
    "openai>=1.82.1",
    "pokebase>=1.4.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
